+++ src/main/java/gregtech/common/tileentities/machines/multi/GT_MetaTileEntity_DieselEngine.java
@@ -356,7 +356,7 @@ public class GT_MetaTileEntity_DieselEngine
                     + EnumChatFormatting.RESET,
             StatCollector.translateToLocal("GT5U.engine.output") + ": "
                 + EnumChatFormatting.RED
-                + GT_Utility.formatNumbers((mEUt * mEfficiency / 10000))
+                + GT_Utility.formatNumbers(((long) mEUt * mEfficiency / 10000))
                 + EnumChatFormatting.RESET
                 + " EU/t",
             StatCollector.translateToLocal("GT5U.engine.consumption") + ": "
